# openapi/openapi.yaml
...
openapi: 3.1.0
components:
  schemas:
    AbonnementDataSeul:
      type: object
      properties:
        id:
          type: string
          description: Identifiant du forfait.
        volumetrie:
          type: string
          description: Quantité de données mobiles.
        debit:
          type: string
          description: Débit de connexion.
        prix:
          type: number
          format: double
          description: Prix du forfait.
        url:
          type: string
          description: URL de la page du forfait sur le site officiel.
        type_forfait:
          type: string
          description: "Type de forfait : internet mobile au volume ou internet mobile\
            \ 4G"
...
    ForfaitBloque:
      description: Forfait bloqué
      type: object
      properties:
        id:
          type: string
          description: Identifiant du forfait bloqué.
        prix:
          type: integer
          format: int32
          description: Prix du forfait bloqué.
        credit:
          type: integer
          format: int32
          description: Montant de crédit du kit.
        sms_offert:
          type: integer
          format: int32
          description: Nombre de SMS offert dans le forfait.
        url:
          type: string
          description: Url de la page du forfait sur le site officiel.
    ForfaitM:
      description: Forfait de la gamme M
      type: object
      properties:
        id:
          type: string
          description: Identifiant du forfait.
        volumetrie:
          type: string
          description: Quantité de données mobiles.
        vocal:
          type: string
          description: Temps d'appel disponible.
        sms:
          type: string
          description: Nombre de SMS disponibles.
        prix:
          type: number
          format: double
          description: Prix du forfait.
        url:
          type: string
          description: URL de la page du forfait sur le site officiel.
    KitPrepaye:
      description: Forfait prépayé
      type: object
      properties:
        id:
          type: string
          description: Identifiant du kit prépayé/recharge liberté.
        credit:
          type: integer
          format: int32
          description: Montant de crédit du kit/recharge.
        prix:
          type: integer
          format: int32
          description: Prix du kit prépayé/recharge liberté.
        sms_offert:
          type: integer
          format: int32
          description: Nombre de SMS inclus dans le kit/recharge.
        duree_validite:
          type: integer
          format: int32
          description: Durée de validité du kit/recharge (en jours).
        url:
          type: string
          description: Url de la page du kit sur le site officiel.
...
paths:
...
  /offres/abonnement-data-seul/{id}:
    get:
      summary: Détails d'un abonnement spécifique
      description: Retourne les détails d'un abonnement spécifique de données en fonction
        de l'ID fourni.
      tags:
      - Abonnement-data-seul
      parameters:
      - description: ID de l'abonnement
        example: IMV-10
        name: id
        required: true
        in: path
        schema:
          type: string
      responses:
        "200":
          description: Détails de l'abonnement
          content:
            application/json:
              examples:
                Exemple de réponse:
                  value:
                    id: IMV-10
                    volumetrie: 1 Mo
                    debit: 256 Ko/s
                    prix: 530
                    url: https://www.opt.nc/sites/serviciel/files/media/file/FI_Internet%20Mobile%20au%20Volume.pdf
                    type_forfait: IMV
              schema:
                $ref: "#/components/schemas/AbonnementDataSeul"
        "404":
          description: Abonnement non trouvé
...
  /offres/forfait-bloque/{id}:
    get:
      summary: Détails d'un forfait bloqué par ID
      description: Retourne les détails d'un forfait bloqué spécifique en fonction
        de l'ID fourni.
      tags:
      - Forfaits bloqués
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Détails du forfait bloqué
          content:
            application/json:
              examples:
                Exemple de réponse:
                  value:
                    id: forfait-bloque-1000
                    credit: 1000
                    prix: 1060
                    sms_offert: 20
                    url: https://www.opt.nc/particuliers/mobile/quel-forfait-choisir/forfait-bloque-1000
              schema:
                $ref: "#/components/schemas/ForfaitBloque"
        "404":
          description: Forfait bloqué non trouvé
...
  /offres/forfait-m/{id}:
    get:
      summary: Détails d'un forfait spécifique de la gamme forfait-m
      description: Retourne les détails d'un forfait spécifique de la gamme forfait-m
        en fonction de l'ID fourni.
      tags:
      - Forfait-m
      parameters:
      - description: ID du forfait
        example: forfait-m-1
        name: id
        required: true
        in: path
        schema:
          type: string
      responses:
        "200":
          description: Détails du forfait
          content:
            application/json:
              examples:
                Exemple de réponse:
                  value:
                    id: forfait-m-1
                    volumetrie: 1 Go
                    vocal: 1 H
                    sms: Illimité
                    prix: 1000.0
                    url: https://www.opt.nc/particuliers/mobile/quel-forfait-choisir/forfait-m-1-go
              schema:
                $ref: "#/components/schemas/ForfaitM"
        "404":
          description: Forfait non trouvé
...
  /offres/prepaye/{id}:
    get:
      summary: Détails d'un forfait prépayé par ID
      description: Retourne les détails d'un forfait prépayé spécifique en fonction
        de l'ID fourni.
      tags:
      - Kit Prépayé
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Détails du forfait prépayé
          content:
            application/json:
              examples:
                Exemple de réponse:
                  value:
                    id: kit-prepaye
                    credit: 3000
                    prix: 6000
                    sms_offert: 0
                    duree_validite: 90
                    url: https://www.opt.nc/particuliers/mobile/quel-forfait-choisir/kit-prepaye-liberte
              schema:
                $ref: "#/components/schemas/KitPrepaye"
        "404":
          description: Forfait prépayé non trouvé
...
info:
  title: forfaits-opt-nc API
  version: 1.0.0-SNAPSHOT
servers:
- url: http://localhost:8080
  description: Auto generated value
- url: http://0.0.0.0:8080
  description: Auto generated value